+++ src/components/sections/FeatureSpotlight.js
